+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/components/richcomponents/ArchiveDropdownIconLaunched.kt
@@ -56,6 +56,13 @@ fun ArchiveDropdownIconLaunched(
                     Text("Associate")
                 })
             }
+
+            DropdownMenuItem(onClick = {
+                operationsLaunchers.openForgetRepository(repositoryURI)
+                isDropdownExpanded = false
+            }, text = {
+                Text("Forget")
+            })
         }
     }
 }